+++ js/data.js
@@ -1,0 +1,4 @@
+const MAX_STRING_LENGTH = 15;
+const str = 'Random String';
+
+export {str, MAX_STRING_LENGTH,};
